Move new method from public API
