+++ hazelcast/src/test/java/com/hazelcast/jet/core/DAGTest.java
@@ -35,7 +35,6 @@ import java.util.Set;
 
 import static com.hazelcast.jet.core.Edge.between;
 import static com.hazelcast.jet.core.Edge.from;
-import static com.hazelcast.jet.core.Vertex.LOCAL_PARALLELISM_USE_DEFAULT;
 import static com.hazelcast.jet.core.processor.Processors.noopP;
 import static java.util.Arrays.asList;
 import static org.hamcrest.CoreMatchers.not;
@@ -385,100 +384,5 @@ public class DAGTest {
 
         // When
         dag.edge(from(b).to(c, 1));
-    }
-
-    @Test
-    public void when_loweringDownParallelismBetweenTwoDefaultParallelismVertices_then_parallelismIsChanged() {
-        DAG dag = new DAG();
-        Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
-        Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
-        dag.edge(between(a, b));
-        dag.lowerDownParallelism(16);
-
-        assertEquals(a.getLocalParallelism(), 4);
-        assertEquals(b.getLocalParallelism(), 4);
-    }
-
-    @Test
-    public void when_loweringDownParallelismBetweenThreeDefaultParallelismVertices_then_parallelismIsChanged() {
-        DAG dag = new DAG();
-        Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
-        Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
-        Vertex c = dag.newVertex("c", PROCESSOR_SUPPLIER);
-        dag.edge(between(a, b));
-        dag.edge(between(b, c));
-        dag.lowerDownParallelism(16);
-
-        assertEquals(a.getLocalParallelism(), 4);
-        assertEquals(b.getLocalParallelism(), 4);
-        assertEquals(c.getLocalParallelism(), 4);
-    }
-
-    @Test
-    public void when_loweringDownParallelismBetweenOneDefaultParallelismVertex_then_parallelismIsNotChanged() {
-        DAG dag = new DAG();
-        Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
-        Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
-
-        a.localParallelism(1);
-        dag.edge(between(a, b));
-        dag.lowerDownParallelism(16);
-
-        assertEquals(a.getLocalParallelism(), 1);
-        assertEquals(b.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
-    }
-
-    @Test
-    public void when_loweringDownParallelismWithSingleCooperativeThreadCount_then_parallelismIsNotChanged() {
-        DAG dag = new DAG();
-        Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
-        Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
-
-        dag.edge(between(a, b));
-        dag.lowerDownParallelism(1);
-
-        assertEquals(a.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
-        assertEquals(b.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
-    }
-
-    @Test
-    public void when_loweringDownParallelismOnLocalEdge_then_edgeIsIsolated() {
-        DAG dag = new DAG();
-        Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
-        Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
-        dag.edge(between(a, b));
-        dag.lowerDownParallelism(16);
-
-        assertEquals(Edge.RoutingPolicy.ISOLATED, dag.getOutboundEdges(a.getName()).get(0).getRoutingPolicy());
-    }
-
-    @Test
-    public void when_loweringDownParallelismOnDistributedEdge_then_parallelismIsNotChanged() {
-        DAG dag = new DAG();
-        Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
-        Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
-        dag.edge(between(a, b));
-        Edge edge = dag.getOutboundEdges(a.getName()).get(0);
-        edge.distributed();
-
-        dag.lowerDownParallelism(16);
-
-        assertEquals(a.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
-        assertEquals(b.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
-    }
-
-    @Test
-    public void when_loweringDownParallelismOnDistributedEdge_then_edgeRoutingPolicyIsNotChanged() {
-        DAG dag = new DAG();
-        Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
-        Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
-        dag.edge(between(a, b));
-        Edge edge = dag.getOutboundEdges(a.getName()).get(0);
-        edge.distributed();
-        Edge.RoutingPolicy routingPolicy = edge.getRoutingPolicy();
-
-        dag.lowerDownParallelism(16);
-
-        assertEquals(routingPolicy, edge.getRoutingPolicy());
     }
 }
